requirements.txt
qrcode==6.1
pillow>=7.1.2
autopep8
